SS4.1: Deleting a sermon will now check if the file is still used in another sermon. Only deleting it if it's no longer used.
git-svn-id: http://joomlacode.org/svn/sermon_speaker@428 5d6a53e0-5f49-0410-96cc-992822d07570

+++ SermonSpeaker4.0/com_sermonspeaker/admin/models/sermon.php
@@ -33,6 +33,7 @@ class SermonspeakerModelSermon extends JModelAdmin
 		// Initialise variables.
 		$pks		= (array) $pks;
 		$table		= $this->getTable();
+		$db 		= $this->getDbo();
 
 		JImport('joomla.filesystem.file');
 		// Iterate the items to delete the files.
@@ -40,13 +41,23 @@ class SermonspeakerModelSermon extends JModelAdmin
 			if ($table->load($pk)) {
 				if ($this->canDelete($table)) {
 					if($table->audiofile && file_exists(JPATH_ROOT.$table->audiofile)){
-						if (!JFile::delete(JPATH_ROOT.$table->audiofile)){
-							$this->setError('Could not delete: '.JPATH_ROOT.$table->audiofile);
+						$query = "SELECT count(1) FROM `#__sermon_sermons` \n"
+								."WHERE `audiofile` = '".$table->audiofile."' OR `videofile` = '".$table->audiofile."'";
+						$db->setQuery($query);
+						if($db->loadResult() == 1){
+							if (!JFile::delete(JPATH_ROOT.$table->audiofile)){
+								$this->setError('Could not delete: '.JPATH_ROOT.$table->audiofile);
+							}
 						}
 					}
 					if($table->videofile && file_exists(JPATH_ROOT.$table->videofile)){
-						if (!JFile::delete(JPATH_ROOT.$table->videofile)){
-							$this->setError('Could not delete: '.JPATH_ROOT.$table->videofile);
+						$query = "SELECT count(1) FROM `#__sermon_sermons` \n"
+								."WHERE `audiofile` = '".$table->videofile."' OR `videofile` = '".$table->videofile."'";
+						$db->setQuery($query);
+						if($db->loadResult() == 1){
+							if (!JFile::delete(JPATH_ROOT.$table->videofile)){
+								$this->setError('Could not delete: '.JPATH_ROOT.$table->videofile);
+							}
 						}
 					}
 				}
